[perf] Default Wan VAE decode to bf16 (lossless, faster)
Wan's pipeline configs left vae_precision at the inherited PipelineConfig
default of fp32, while the DiT already runs bf16 and the *same* AutoencoderKLWan
runs bf16 in Cosmos-Predict2.5 and fp16 in Cosmos. Measured on a real Wan latent
(decoded fp32 vs bf16 in one process):
  - MS-SSIM(bf16, fp32) = 0.9999 on the identical latent (no quality cost)
  - ~1.2-1.3x faster VAE decode (~5-10% end-to-end on decode-bound few-step
    models; free quality-wise on full-step)

fp32 here was the inherited default, not a Wan-specific requirement; this aligns
Wan with its sibling configs. Also updates the two legacy wan_*.json configs.

+++ fastvideo/configs/pipelines/wan.py
@@ -47,7 +47,11 @@ class WanT2V480PConfig(PipelineConfig):
 
     # Precision for each component
     precision: str = "bf16"
-    vae_precision: str = "fp32"
+    # bf16 VAE decode is effectively lossless (MS-SSIM 0.9999 vs fp32 on an
+    # identical latent) and faster; the same AutoencoderKLWan already runs bf16
+    # in Cosmos-Predict2.5 and fp16 in Cosmos. fp32 here was just the inherited
+    # PipelineConfig default, not a Wan-specific requirement.
+    vae_precision: str = "bf16"
     text_encoder_precisions: tuple[str, ...] = field(default_factory=lambda: ("fp32", ))
 
     # self-forcing params
